backend/requirements.txt
django==4.2.11
djangorestframework==3.15.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
djangorestframework-simplejwt==5.3.1
python-dotenv==1.0.1
